Update .gitignore for Docker Volumes and Configuration
This commit introduces the following changes to the codebase:

- Added `.env` to the `.gitignore` file
- Added `keycloak-postgres-data` to the `.gitignore` file
- Added `mongo-product-service` to the `.gitignore` file
- Added `postgres-inventory-service` to the `.gitignore` file
- Added `postgres-order-service` to the `.gitignore` file
- Removed `src/main/resources/docker-compose` from the `.gitignore` file

Purpose:
The purpose of this commit is to update the `.gitignore` file to exclude Docker volumes and configuration files from being tracked by Git. This helps to prevent sensitive data or large volume data from being committed to the repository accidentally. Additionally, removing the `src/main/resources/docker-compose` entry allows for better management of Docker Compose files in the project.