Add Keycloak Realm Configuration
This commit introduces the following changes to the codebase:

- Added `realm-export.json` file containing the Keycloak realm configuration for the "spring-boot-shopping-realm"

Purpose:
The purpose of this commit is to include the Keycloak realm configuration file `realm-export.json` in the codebase. This file contains the complete configuration for the "spring-boot-shopping-realm" realm, including user accounts, client applications, roles, and other settings. Adding this file to the codebase allows for easy setup and deployment of the Keycloak realm for the Spring Boot shopping application.