Add Zookeeper and Kafka Services to Docker Compose
This commit introduces the following changes to the codebase:

- Added Zookeeper service configuration to the Docker Compose file
- Added Kafka service configuration to the Docker Compose file
- Set Kafka service to depend on Zookeeper service

Purpose:
The purpose of this commit is to add Zookeeper and Kafka services to the Docker Compose file. Zookeeper is a distributed coordination service required by Kafka, which is a distributed event streaming platform. The Order Service, which depends on Kafka, will be able to interact with the Kafka service defined in this Docker Compose configuration. By incorporating Zookeeper and Kafka into the Docker Compose setup, the entire application stack can be easily deployed and managed in a containerized environment.