Add Docker Compose Configuration for Notification Service
This commit introduces the following changes to the codebase:

- Add a new service configuration for the `notification-service` in the `docker-compose.yaml` file.
- Set the Docker image for the notification service to `aamirxshaikh/notification-service:latest`.
- Set the container name for the notification service to `notification-service`.
- Configure the `SPRING_PROFILES_ACTIVE` environment variable to `docker` for the notification service.
- Make the notification service dependent on the `zipkin`, `kafka`, `discovery-server`, and `api-gateway` services.

Purpose:
The purpose of this commit is to include the `notification-service` in the Docker Compose configuration. By adding the service configuration to the `docker-compose.yaml` file, the notification service can be easily deployed and managed alongside the other microservices in the application. The commit configures the necessary settings, such as the Docker image, container name, active profile, and service dependencies, to ensure the proper functioning of the notification service within the Docker Compose environment.